+++ packages/@leight/job-server/src/sdk/ServerBaseSource.ts
@@ -1,28 +1,21 @@
 /**
- Source code containing implementation of Server-side Source for Job, TRPC router part (if no disabled) and
- some other cool stuff.
+ Base Source contains default implementation of Source for entity Job. This could be used for further extensions,
+ also default export uses this as a parent class.
  */
 import {
-    type IContainer,
-    ServiceContext
-}                            from "@leight/container";
-import {
     $JobSource,
-    type IJobSource,
-    type IJobSourceSchema,
-    JobQuerySchema
-}                            from "@leight/job";
+    type IJobSourceSchema
+}                       from "@leight/job";
 import {
     $PrismaClient,
     type PrismaClient
-}                            from "@leight/prisma";
-import {withCursor}          from "@leight/query";
+}                       from "@leight/prisma";
+import {withCursor}     from "@leight/query";
 import {
     type ISource,
     withUpsert
-}                            from "@leight/source";
-import {AbstractSource}      from "@leight/source-server";
-import {withSourceProcedure} from "@leight/trpc-source-server";
+}                       from "@leight/source";
+import {AbstractSource} from "@leight/source-server";
 
 type IEntity = IJobSourceSchema["Entity"];
 type IQuery = IJobSourceSchema["Query"];
@@ -62,12 +55,3 @@ export class JobBaseSource extends AbstractSource<IJobSourceSchema> {
         return this.prismaClient.job;
     }
 }
-
-export class JobSource extends JobBaseSource implements IJobSource {
-}
-
-export const JobSourceContext   = (container: IContainer) => new ServiceContext<IJobSource>(container, $JobSource);
-export const JobSourceProcedure = withSourceProcedure<IJobSourceSchema>({
-    source: $JobSource,
-    schema: JobQuerySchema,
-});