+++ packages/@leight/job-server/src/sdk/ServerSource.ts
@@ -3,3 +3,8 @@
 
  So, please, DO NOT modify this file as it would get re-generated and you would be f*cked up.
  */
+import {type IJobSource} from "@leight/job";
+import {JobBaseSource}   from "./ServerBaseSource";
+
+export class JobSource extends JobBaseSource implements IJobSource {
+}