+++ packages/@leight/sdk/src/api/index.ts
@@ -22,6 +22,17 @@ export interface IGenerator<TParams = void> {
     (props: ITemplate<TParams>): Promise<void>;
 }
 
+export interface ISdkGeneratorProps {
+    /**
+     * Package name where the SDK is generated; if not provided, package.json is resolved.
+     */
+    packageName?: string;
+    /**
+     * Target folder of the managed source code; it should be dedicated to the generator only.
+     */
+    folder?: string;
+}
+
 export interface ISdkGenerator {
     (): Promise<void>;
 }