+++ packages/@leight/sdk/src/utils/withClientSourceGenerators.ts
@@ -1,50 +1,30 @@
-import {resolvePackageJson} from "@leight/utils-server";
+import {resolvePackageJson}      from "@leight/utils-server";
+import {type ISdkGeneratorProps} from "../api";
 import {
-    generatorClientContext,
     generatorClientSource,
-    type IGeneratorClientContextParams,
-    type IGeneratorClientSourceParams
-}                           from "../generator";
+    type IGeneratorClientSourceParams,
+}                                from "../generator";
 
 export type IWithClientSourceGeneratorsProps =
-    IGeneratorClientContextParams
-    &
     IGeneratorClientSourceParams
-    & {
-        packageName?: string;
-        sdk?: string;
-    }
+    &
+    ISdkGeneratorProps;
 
 export const withClientSourceGenerators = (
     {
         packageName = resolvePackageJson().name,
-        sdk = "src/sdk",
-        packages,
-        trpc,
-        entity,
+        folder = "src/sdk",
+        ...params
     }: IWithClientSourceGeneratorsProps) => {
     if (!packageName) {
         throw new Error("Cannot resolve packageName");
     }
     return [
-        async () => generatorClientContext({
-            packageName,
-            folder: `${sdk}/client-context.ts`,
-            barrel: true,
-            params: {
-                entity,
-                packages,
-            },
-        }),
         async () => generatorClientSource({
             packageName,
+            folder,
             barrel: true,
-            folder: `${sdk}/client-source.tsx`,
-            params: {
-                entity,
-                trpc,
-                packages,
-            }
+            params,
         }),
     ];
 };