+++ packages/@leight/sdk/src/utils/withServerSourceGenerator.ts
@@ -1,41 +1,30 @@
-import {resolvePackageJson} from "@leight/utils-server";
+import {resolvePackageJson}      from "@leight/utils-server";
+import {type ISdkGeneratorProps} from "../api";
 import {
-    generatorServerSource,
-    type IGeneratorServerSourceParams
-}                           from "../generator";
+    generatorServer,
+    type IGeneratorServerParams
+}                                from "../generator";
 
 export type IWithServerSourceGeneratorsProps =
-    IGeneratorServerSourceParams
-    & {
-        packageName?: string;
-        sdk?: string;
-    }
+    IGeneratorServerParams
+    &
+    ISdkGeneratorProps;
 
 export const withServerSourceGenerators = (
     {
         packageName = resolvePackageJson().name,
-        sdk = "src/sdk",
-        entity,
-        packages,
-        prisma,
-        disabled,
-        sourceEx,
+        folder = "src/sdk",
+        ...params
     }: IWithServerSourceGeneratorsProps) => {
     if (!packageName) {
         throw new Error("Cannot resolve packageName");
     }
     return [
-        async () => generatorServerSource({
+        async () => generatorServer({
             packageName,
+            folder,
             barrel: true,
-            folder: `${sdk}/server-source.tsx`,
-            params: {
-                entity,
-                packages,
-                prisma,
-                disabled,
-                sourceEx,
-            }
+            params,
         }),
     ];
 };