+++ packages/@leight/generator-server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/generator-server
 
+## 0.5.3
+
+### Patch Changes
+
+- Refactored generator, improved a lot of stuff
+
 ## 0.5.2
 
 ### Patch Changes